[IMP] core,website: move `MockRequest` in `tests.common`
The `MockRequest` context manager is only used in tests.
Move it to `odoo.tests.common.py`.

task-4822364
