+++ addons/test_website/tests/test_views_during_module_operation.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from odoo.addons.website.tools import MockRequest
+from odoo.tests.common import MockRequest
 from odoo.tests import standalone
