+++ addons/website/tests/test_lang_url.py
@@ -4,7 +4,7 @@ import json
 import lxml.html
 from urllib.parse import urlparse
 
-from odoo.addons.website.tools import MockRequest
+from odoo.tests.common import MockRequest
 from odoo.tests import HttpCase, tagged
