+++ addons/website/tests/test_qweb.py
@@ -2,7 +2,7 @@
 
 from odoo import http
 from odoo.addons.base.tests.common import TransactionCaseWithUserDemo
-from odoo.addons.website.tools import MockRequest
+from odoo.tests.common import MockRequest
 from odoo.tests.common import TransactionCase
