+++ addons/website/tests/test_redirect.py
@@ -1,6 +1,6 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from odoo.addons.website.tools import MockRequest
+from odoo.tests.common import MockRequest
 from odoo.exceptions import ValidationError
 from odoo.tests import TransactionCase, tagged
