+++ addons/website_project/tests/test_project_portal_access.py
@@ -6,7 +6,7 @@ from odoo.tests import HttpCase
 
 from odoo.addons.mail.controllers.thread import ThreadController
 from odoo.addons.project.tests.test_project_sharing import TestProjectSharingCommon
-from odoo.addons.website.tools import MockRequest
+from odoo.tests.common import MockRequest
 
 
 class TestProjectPortalAccess(TestProjectSharingCommon, HttpCase):
